+++ src/fmpz_mod_poly/hamming_weight.c
@@ -1,0 +1,19 @@
+/*
+    Copyright (C) 2023 Albin Ahlbäck
+
+    This file is part of FLINT.
+
+    FLINT is free software: you can redistribute it and/or modify it under
+    the terms of the GNU Lesser General Public License (LGPL) as published
+    by the Free Software Foundation; either version 2.1 of the License, or
+    (at your option) any later version.  See <https://www.gnu.org/licenses/>.
+*/
+
+#include "fmpz_poly.h"
+#include "fmpz_mod_poly.h"
+
+slong
+fmpz_mod_poly_hamming_weight(const fmpz_mod_poly_t A, const fmpz_mod_ctx_t ctx)
+{
+    return _fmpz_poly_hamming_weight(A->coeffs, A->length);
+}
