+++ src/fmpz_mod_poly/truncate.c
@@ -1,0 +1,26 @@
+/*
+    Copyright (C) 2023 Albin Ahlbäck
+
+    This file is part of FLINT.
+
+    FLINT is free software: you can redistribute it and/or modify it under
+    the terms of the GNU Lesser General Public License (LGPL) as published
+    by the Free Software Foundation; either version 2.1 of the License, or
+    (at your option) any later version.  See <https://www.gnu.org/licenses/>.
+*/
+
+#include "fmpz.h"
+#include "fmpz_mod_poly.h"
+
+void fmpz_mod_poly_truncate(fmpz_mod_poly_t poly, slong len, const fmpz_mod_ctx_t ctx)
+{
+    if (poly->length > len)
+    {
+        slong i;
+
+        for (i = len; i < poly->length; i++)
+            _fmpz_demote(poly->coeffs + i);
+        poly->length = len;
+        _fmpz_mod_poly_normalise(poly);
+    }
+}
